+++ packages/adminui-vue/components/controls/USelectEntity.vue
@@ -106,7 +106,7 @@
             style="margin: 5px"
             @click="showMore"
           >
-            {{ $ut('more') }}
+            {{ $ut('USelectEntity.dropdown.moreButton') }}
           </el-button>
         </el-row>
       </div>