+++ packages/adminui-vue/CHANGELOG.md
@@ -7,8 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
  - `UMasterDetailView`: sort order saved in the localStorage for each shortcut
- - Excel export from UTableEntity adds a total row with SUM formula for a numeric columnus
- - `UToolbar.showAccessRights`: added `parentContext.instanceID` (ID of the ubstance we adds rigths for) to the `addNew` command
+ - Excel export from UTableEntity adds a total row with SUM formula for a numeric columns
+ - `UToolbar.showAccessRights`: added `parentContext.instanceID` (ID of the instance we add rigths for) to the `addNew` command
+ - The `processing` vuex module now has `PURGE_COLLECTION_ITEM` mutation.  It acts like `DELETE_COLLECTION_ITEM`,
+   but does NOT track deletion.  This is useful for collection items, which depends on other collections and ARE
+   deleted by cascade.
  
 ### Changed
  - `vue` updated 2.6.10 -> 2.6.12; `vue-template-compiler` updated 2.6.10 -> 2.6.12