+++ packages/adminui-vue/CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
  - `UTableEntity`: action `show summary` in dropdown menu which shows summary data for each column with type Number.
-   Also added slot for summary action - `toolbarDropdownSummary`
+   Also, added slot for summary action - `toolbarDropdownSummary`
  - `UMasterDetailView`: sort order saved in the localStorage for each shortcut
  - Excel export from UTableEntity adds a total row with SUM formula for a numeric columns
  - `UToolbar.showAccessRights`: added `parentContext.instanceID` (ID of the instance we add rigths for) to the `addNew` command
@@ -36,7 +36,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  - `mountContainer` provide stubs `$formServices.setTitle` & `$formServices.close` as `mountTab` and `mountModal` do.
   This allows form to be mounted inside preview area without changing form code (in case form use these methods).
  - `UTableEntity` update lookup data only in case of updated record match current repository whereList conditions.
-  This prevents adding unexpected values to lookups in opened forms 
+  This prevents adding unexpected values to lookups in opened forms.
 
 ## [1.13.4] - 2020-09-11
 ### Added